+++ tests/test_all_angle.py
@@ -26,7 +26,7 @@ def test_all_angle_bundle(LAYER: kf.LayerEnum) -> None:
         ae_rad = np.deg2rad(ae)
         start_ports.append(
             vc.create_port(
-                name="s0",
+                name=f"s{i}",
                 dcplx_trans=kf.kdb.DCplxTrans(
                     1, a, False, -500 + r * np.cos(a_rad), -100 + r * np.sin(a_rad)
                 ),
@@ -36,22 +36,26 @@ def test_all_angle_bundle(LAYER: kf.LayerEnum) -> None:
         )
         end_ports.append(
             vc.create_port(
-                name="s0",
+                name=f"s{i + _l}",
                 dcplx_trans=kf.kdb.DCplxTrans(
-                    1, ae, False, 1510 + r * np.cos(ae_rad), 1410 + r * np.sin(ae_rad)
+                    1, ae, False, 2510 + r * np.cos(ae_rad), 2410 + r * np.sin(ae_rad)
                 ),
                 layer=LAYER.WG,
                 dwidth=1,
             )
         )
-
+    backbone = [
+        kf.kdb.DPoint(750, -550),
+        kf.kdb.DPoint(1000, 550),
+        kf.kdb.DPoint(1000, 1200),
+        kf.kdb.DPoint(2800, 1950),
+    ]
     kf.routing.aa.optical.route_bundle(
         vc,
         start_ports=start_ports,
         end_ports=end_ports,
-        backbone=[kf.kdb.DPoint(550, 550), kf.kdb.DPoint(950, 950)],
-        spacings=[randint(1, 5) for _ in range(_l)],
-        radius_estimate=lambda angle: 60,
+        backbone=backbone,
+        separation=[randint(1, 5) for _ in range(_l)],
         straight_factory=sf,
         bend_factory=bf,
     )